Move API routes
Moves API routes to separte module in routes for modular routing.
